Add files via upload
version 1 of Geocoding integration using Google Maps. New class Geocoding with a function that validates a postcode and returns true or false depending on if that postcode can be found in Google Maps. 

Further improvements soon to be made. Accepting different types of addresses, not just poscode and interpreting Json formatted responses to be used later in the application if required